Upgrade to use NCS v3.2.0

+++ src/ble_provisioning.c
@@ -94,7 +94,7 @@ static const struct bt_data sd[] = {
 static struct k_work_delayable update_adv_param_work;
 static struct k_work_delayable update_adv_data_work;
 
-static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 				    struct net_if *iface)
 {
 	struct wifi_iface_status status = {0};
